refactor(api): move CertificationRescored from session management to shared
